server_python/requirements.txt
fastapi
uvicorn[standard]
motor
python-dotenv
bcrypt
pyjwt
pydantic[email]
python-multipart
